migrate logging into seperate mod
